A capture of various trials for print-preview and print.
The WPFPrintEngine test has some nice capabilities but introduces MVVM, is difficult to get the "full" chart, doesn't have per-page margins, etc. The necessary rework to may be difficult.

The VisualPrint test does a "better" job of paging, and should be easier to evolve to be more useful. However, text sometimes prints corrupted, it feels slower to print/preview, and work will be required to support orientation, etc.