Introduce a simple MVC scenario: the FamilyView raises specific events, which are handled in the MainWindow by invoking methods in the Details control. This allows the FamilyView to request that the details pane react without introducing a strong coupling between the FamilyView and the Details pane.
Events are raised for Add Father, Add Mother, View Marriage, Add Son, Add Daughter, and Add Spouse.

There are a couple of 'cobbles' in place: a. the current mechanism for passing custom arguments with the events is a little bit of a hack; b. the "Add relationship" sub-pane isn't really designed for adding to the non-current person.