Snapshot of a work-in-progress. Adding support for more events/facts from GEDCOM and more event/fact details.
o Created a new 'FactEditor' control. Allows the user to see/edit all supported details about a fact/event.
o Added a button to the fact display in the details panel. This button takes the user to the 'FactEditor' control.
o Importing more fact/event properties from GEDCOM.

The FactEditor is incomplete. The save to Familyx support is untested. Export to GEDCOM is untested. The button to invoke the FactEditor has not been added to Events.